@@ -1,56 +1,76 @@
 {
+    "private": true,
     "name": "raweb",
+    "type": "module",
     "scripts": {
         "dev": "vite",
-        "build": "vite build",
+        "build": "vite build && vite build --ssr",
         "lint": "npm run lint:eslint",
         "lint:eslint": "eslint public/js resources/js postcss.config.js tailwind.config.js vite.config.ts --cache",
         "fix": "npm run fix:eslint",
         "fix:eslint": "eslint public/js resources/js postcss.config.js tailwind.config.js vite.config.ts --fix",
-        "test": "vitest run",
+        "test": "vitest run --reporter=dot",
         "test:watch": "vitest",
-        "analyze": "cross-env ANALYZE=true npm run prod",
-        "apidoc": "npm run apidoc:connect",
-        "apidoc:connect": "apidoc -i src/Connect/ -c docs/src/connect-api -t docs/src/connect-api/template -o docs/dist/api/connect"
+        "analyze": "cross-env ANALYZE=true npm run prod"
     },
     "dependencies": {
         "@floating-ui/dom": "^1.5.1",
+        "@radix-ui/react-navigation-menu": "^1.1.4",
+        "@radix-ui/react-slot": "^1.0.2",
+        "class-variance-authority": "^0.7.0",
+        "clsx": "^2.1.1",
+        "dayjs": "^1.11.11",
         "js-cookie": "^3.0.5",
         "linkify-html": "^4.1.3",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "react-icons": "^5.2.1",
+        "tailwindcss-animate": "^1.0.7",
         "ua-parser-js": "^1.0.35"
     },
     "devDependencies": {
-        "@defstudio/vite-livewire-plugin": "^1.1.0",
+        "@faker-js/faker": "^8.4.1",
+        "@inertiajs/react": "^1.2.0",
         "@tailwindcss/aspect-ratio": "^0.4.2",
         "@tailwindcss/forms": "^0.5.3",
         "@tailwindcss/typography": "^0.5.0",
-        "@testing-library/dom": "^9.3.1",
-        "@testing-library/jest-dom": "^5.16.5",
-        "@testing-library/user-event": "^14.4.3",
+        "@testing-library/dom": "^10.1.0",
+        "@testing-library/jest-dom": "^6.4.6",
+        "@testing-library/react": "^16.0.0",
+        "@testing-library/user-event": "^14.5.2",
         "@types/alpinejs": "^3.7.2",
         "@types/alpinejs__focus": "^3.10.0",
         "@types/js-cookie": "^3.0.3",
+        "@types/react": "^18.3.3",
+        "@types/react-dom": "^18.3.0",
         "@types/ua-parser-js": "^0.7.36",
         "@typescript-eslint/eslint-plugin": "^5.60.1",
         "@typescript-eslint/parser": "^5.60.1",
+        "@vitejs/plugin-react": "^4.3.1",
         "alpinejs": "^3.12.2",
         "autoprefixer": "^10.4.14",
         "eslint": "^8.44.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-import-resolver-typescript": "^3.5.5",
+        "eslint-plugin-jsx-a11y": "^6.8.0",
         "eslint-plugin-prettier": "^5.1.3",
+        "eslint-plugin-react": "^7.34.2",
+        "eslint-plugin-react-hooks": "^4.6.2",
         "eslint-plugin-simple-import-sort": "^12.0.0",
+        "eslint-plugin-testing-library": "^6.2.2",
+        "eslint-plugin-unicorn": "^54.0.0",
         "jsdom": "^22.1.0",
         "laravel-echo": "^1.15.1",
-        "laravel-vite-plugin": "^0.8.0",
+        "laravel-vite-plugin": "^1.0.4",
         "postcss": "^8.4.31",
         "postcss-nesting": "^12.1.5",
         "prettier": "3.2.5",
         "pusher-js": "^8.2.0",
+        "tailwind-merge": "^2.3.0",
         "tailwindcss": "^3.4.3",
-        "typescript": "^4.9.5",
-        "vite": "^4.5.2",
-        "vitest": "^0.34.6"
+        "typescript": "^5.4.5",
+        "vite": "^5.2.13",
+        "vitest": "^1.6.0"
     },
     "engines": {
         "node": "20.x",
