@@ -1,14 +1,23 @@
+@props([
+    'page' => null,
+])
+
 <head prefix="og: http://ogp.me/ns# retroachievements: http://ogp.me/ns/apps/retroachievements#">
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>{{ (!empty($pageTitle) ? $pageTitle . ' · ' : '') . config('app.name') }}</title>
+
+    {{-- Inertia.js pages cannot override an existing title or meta description. --}}
+    @if (empty($page))
+        <title>{{ (!empty($pageTitle) ? $pageTitle . ' · ' : '') . config('app.name') }}</title>
+        <meta name="description" content="{{ $pageDescription ?? __('app.description') }}">
+    @endif
+
     <link rel="icon" type="image/png" href="{{ asset(app()->environment('local', 'stage') ? 'assets/images/favicon-gray.webp' : 'assets/images/favicon.webp') }}">
     <link rel="preload" as="image" importance="high" href="{{ asset('assets/images/ra-icon.webp') }}">
     <link rel="image_src" href="{{ asset('assets/images/ra-icon.webp') }}">
     <meta name="csrf-token" content="{{ csrf_token() }}">
     <meta name="copyright" content="Copyright 2014-{{ date('Y') }}">
-    <meta name="description" content="{{ $pageDescription ?? __('app.description') }}">
     <meta name="keywords" content="games, achievements, retro, emulator">
     <meta name="format-detection" content="telephone=no">
     @if (config('services.facebook.client_id'))
@@ -32,7 +41,13 @@
     {{-- TODO replace with ESM imports, Alpine, tailwind --}}
     <link rel="stylesheet" href="https://ajax.googleapis.com/ajax/libs/jqueryui/1.8.23/themes/sunny/jquery-ui.css">
 
-    @vite(['resources/js/app.ts', 'resources/css/app.css'], config('vite.build_path'))
+    @routes
+    @viteReactRefresh
+    @vite(['resources/js/tall-stack/app.ts', 'resources/css/app.css', 'resources/js/app.tsx'], config('vite.build_path'))
+    @if (!empty($page))
+        @inertiaHead
+        @vite(['resources/js/app.tsx', "resources/js/pages/{$page['component']}.tsx"], config('vite.build_path'))
+    @endif
 
     <livewire:styles />
 
