@@ -1,12 +1,13 @@
 /// <reference types="vitest" />
 
+import react from '@vitejs/plugin-react';
 import { existsSync, readFileSync } from 'fs';
 import laravel from 'laravel-vite-plugin';
 import { homedir } from 'os';
 import { resolve } from 'path';
 import { defineConfig, loadEnv } from 'vite';
 
-export default defineConfig(({ mode }) => {
+export default defineConfig(({ mode, isSsrBuild }) => {
   const env = loadEnv(mode, process.cwd(), '');
 
   if (!env.VITE_BUILD_PATH) {
@@ -18,30 +19,40 @@ export default defineConfig(({ mode }) => {
   }
 
   return {
+    // Required for SSR assets to load properly
+    base: '/assets/build',
+
     // https://vitejs.dev/config/#build-options
     build: {
-      outDir: `public/${env.VITE_BUILD_PATH}`,
+      outDir: isSsrBuild ? `bootstrap/ssr` : `public/${env.VITE_BUILD_PATH}`,
       assetsDir: '',
       assetsInlineLimit: 4096,
     },
+
     // https://vitejs.dev/config/#plugins
     plugins: [
       laravel({
-        input: ['resources/css/app.css', 'resources/js/app.ts'],
+        input: ['resources/css/app.css', 'resources/js/tall-stack/app.ts', 'resources/js/app.tsx'],
+        ssr: 'resources/js/ssr.tsx',
         refresh: ['resources/views/**'],
       }),
+      react(),
     ],
+
     resolve: {
       alias: {
         '@': resolve(__dirname, './resources/js'),
         livewire: resolve(__dirname, './vendor/livewire/livewire/dist/livewire.esm'),
       },
     },
+
     test: {
       environment: 'jsdom',
       setupFiles: 'resources/js/setupTests.ts',
-      include: ['resources/js/**/*.{test,spec}.ts'],
+      include: ['resources/js/**/*.{test,spec}.{ts,tsx}'],
+      globals: true,
     },
+
     // @ see https://vitejs.dev/config/#server-options
     server: detectServerConfig(env),
   };
